move synchronize logic to consensus
